Add 'admin' field to okapi_consumers
Beta. Administrators can use this field to mark a chosen Consumers as
"administrative". Being "Administrative Consumer" allows him to skip
some of the safety limits which are imposed on all other Consumers (in
order to enforce efficient use).

It should be set only for trusted and competent Consumers. If used
incorrectly, administrative keys may kill OKAPI!